feat: 감정 기록 삭제 훅 추가
